SAST-0: retire the last 2 long-as-pointer truncations (savetemp + savegame reloc)
savetemp(char*,long daptr,...) truncated its pointer arg on x64 -> callers pass
waloff[MAXTILES-1] (intptr_t tile-data pointer), so the temp-save wrote from a truncated/
bad pointer (the 'waloff bad pointer' class). Param + prototype -> intptr_t. Also the
savegame T-value reloc base j=(long)&script[0] -> (intptr_t) (sibling of the migration's
other reloc casts). SAST HIGH 2 -> 0: the entire long-as-pointer cluster (the original 23,
== the deferred E1 work) is now fully retired. Full suite 2066 green; 5/5 flags.

+++ source/MENUES.C
@@ -120,7 +120,7 @@ void cmenu(short cm)
 }
 
 
-void savetemp(char *fn,long daptr,long dasiz)
+void savetemp(char *fn,intptr_t daptr,long dasiz)
 {
     int fp;
 
@@ -996,7 +996,7 @@ saveplayer(signed char spot)
     for(i=0;i<MAXSPRITES;i++)
     {
         if(actorscrptr[PN] == 0) continue;
-        j = (long)&script[0];
+        j = (intptr_t)&script[0];
 
         if(scriptptrs[i]&1)
             T2 += j;
